# Run blocking code on the compute pool more carefully