--- a/core/shared/src/main/scala/cats/effect/IOFiber.scala
+++ b/core/shared/src/main/scala/cats/effect/IOFiber.scala
@@ -943,19 +943,33 @@ private final class IOFiber[A](
           if (cur.hint eq IOFiber.TypeBlocking) {
             val ec = currentCtx
             if (ec.isInstanceOf[WorkStealingThreadPool]) {
-              var error: Throwable = null
-              val r =
-                try {
-                  scala.concurrent.blocking(cur.thunk())
-                } catch {
-                  case NonFatal(t) =>
-                    error = t
-                  case t: Throwable =>
-                    onFatalFailure(t)
+              val wstp = ec.asInstanceOf[WorkStealingThreadPool]
+              if (wstp.canExecuteBlockingCode()) {
+                var error: Throwable = null
+                val r =
+                  try {
+                    scala.concurrent.blocking(cur.thunk())
+                  } catch {
+                    case NonFatal(t) =>
+                      error = t
+                    case t: Throwable =>
+                      onFatalFailure(t)
+                  }
+
+                val next = if (error eq null) succeeded(r, 0) else failed(error, 0)
+                runLoop(next, nextCancelation, nextAutoCede)
+              } else {
+                resumeTag = BlockingR
+                resumeIO = cur
+
+                if (isStackTracing) {
+                  val handle = monitor()
+                  objectState.push(handle)
                 }
 
-              val next = if (error eq null) succeeded(r, 0) else failed(error, 0)
-              runLoop(next, nextCancelation, nextAutoCede)
+                val ec = runtime.blocking
+                scheduleOnForeignEC(ec, this)
+              }
             } else {
               resumeTag = BlockingR
               resumeIO = cur