--- a/core/shared/src/main/scala/cats/effect/IOFiber.scala
+++ b/core/shared/src/main/scala/cats/effect/IOFiber.scala
@@ -959,28 +959,10 @@ private final class IOFiber[A](
                 val next = if (error eq null) succeeded(r, 0) else failed(error, 0)
                 runLoop(next, nextCancelation, nextAutoCede)
               } else {
-                resumeTag = BlockingR
-                resumeIO = cur
-
-                if (isStackTracing) {
-                  val handle = monitor()
-                  objectState.push(handle)
-                }
-
-                val ec = runtime.blocking
-                scheduleOnForeignEC(ec, this)
+                blockingFallback(cur)
               }
             } else {
-              resumeTag = BlockingR
-              resumeIO = cur
-
-              if (isStackTracing) {
-                val handle = monitor()
-                objectState.push(handle)
-              }
-
-              val ec = runtime.blocking
-              scheduleOnForeignEC(ec, this)
+              blockingFallback(cur)
             }
           } else {
             runLoop(interruptibleImpl(cur), nextCancelation, nextAutoCede)
@@ -997,6 +979,19 @@ private final class IOFiber[A](
           runLoop(succeeded(Trace(tracingEvents), 0), nextCancelation, nextAutoCede)
       }
     }
+  }
+
+  private[this] def blockingFallback(cur: Blocking[Any]): Unit = {
+    resumeTag = BlockingR
+    resumeIO = cur
+
+    if (isStackTracing) {
+      val handle = monitor()
+      objectState.push(handle)
+    }
+
+    val ec = runtime.blocking
+    scheduleOnForeignEC(ec, this)
   }
 
   /*