Only run blocking code if the worker thread matches the compute runtime.