@vasilmkd 2022-01-08
Benchmark runs:

`Original PR`:
```scala
Benchmark                        (size)   Mode  Cnt     Score     Error  Units
BlockingBenchmark.blockThenCede   10000  thrpt   20     6.229 ±   0.299  ops/s
BlockingBenchmark.coarse          10000  thrpt   20  3565.953 ± 105.111  ops/s
BlockingBenchmark.fine            10000  thrpt   20   758.839 ±  19.670  ops/s
BlockingBenchmark.nested          10000  thrpt   20   729.449 ±  12.753  ops/s
```

`This PR`:
```scala
Benchmark                        (size)   Mode  Cnt     Score     Error  Units
BlockingBenchmark.blockThenCede   10000  thrpt   20     6.481 ±   0.274  ops/s
BlockingBenchmark.coarse          10000  thrpt   20  3481.849 ± 111.972  ops/s
BlockingBenchmark.fine            10000  thrpt   20   771.115 ±   7.107  ops/s
BlockingBenchmark.nested          10000  thrpt   20   752.231 ±   8.088  ops/s
```

No issues. We can merge.